nats_mq_poc/post_svc/requirements.txt
asgiref==3.5.2
asyncio-nats-client==0.11.5
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
greenlet==1.1.3
greenletio==0.9.0
h11==0.14.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
nats-py==2.1.7
protobuf==4.21.6
six==1.16.0
uvicorn==0.18.3
Werkzeug==2.2.2
zipp==3.8.1
